rename: domain/RestaurantRegister 도메인 객체명을 명확히 하기 위해  RestaurantRegistry로 변경
