+++ src/domain/RestaurantRegistry.ts
