+++ src/domain/index.ts
@@ -1,3 +1,3 @@
 export { default as RestaurantDataProvider } from './RestaurantDataProvider';
 
-export { default as RestaurantRegister } from './RestaurantRegister';
+export { default as RestaurantRegistry } from './RestaurantRegistry';
